fix: footer 전체 레이아웃에 추가 및 footer 라우터 추가, 챗봇 userLayout에 추가
