+++ triangle/src/layouts/DescriptionLayout/DescriptionLayout.jsx
@@ -1,10 +1,12 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
+import { Footer } from "../../components/Footer/Footer";
 
 function DescriptionLayout(){
     return(
         <div>
             <Outlet/>
+            <Footer/>
         </div>
     )
 }
