+++ triangle/src/layouts/UserLayout/UserLayout.css
@@ -3,6 +3,7 @@
     flex-direction: column;
     height: 100%;
     width: 100%;
+    min-height: 100vh;
     min-width: 32rem;
     background-color: white
 }
