+++ triangle/src/layouts/UserLayout/UserLayout.jsx
@@ -2,6 +2,8 @@ import React from "react";
 import { Outlet } from "react-router-dom";
 import './UserLayout.css'
 import TopNavbar from '../../components/TopNavbar'
+import { Footer } from "../../components/Footer/Footer";
+import { ChatBot } from "../../components/ChatBot/ChatBot";
 
 function UserLayout(){
     return(
@@ -9,9 +11,10 @@ function UserLayout(){
             <div className="UserLayoutContainer">
                 <TopNavbar/>
                 <div className="UserContentContainer">
+                    <ChatBot/>
                     <Outlet/>
                 </div>
-                
+                <Footer/>
             </div>
         </>
     )
